Added image resize with default resize algorithm

+++ image-processor/src/image.rs
@@ -3,14 +3,14 @@ use wasm_bindgen::prelude::*;
 #[wasm_bindgen]
 #[derive(Clone)]
 pub struct Image {
-    pub width: u16,
-    pub height: u16,
+    pub width: u32,
+    pub height: u32,
     pixels: Vec<u8>,
 }
 
 #[wasm_bindgen]
 impl Image {
-    pub fn new(width: u16, height: u16, pixels: Vec<u8>) -> Self {
+    pub fn new(width: u32, height: u32, pixels: Vec<u8>) -> Self {
         Self{width, height, pixels}
     }
